Phase1
So this app is going to change the blood glucose world forever. In its final phase it will be a gesture activated application on a mobile device or tablet that will send alerts for hypoglycemia or hyperglycemia both of which can end fatally or cause irreversible damage to the body.